+++ src/May/5.12/171.cpp
@@ -13,7 +13,7 @@ class Solution {
   int titleToNumber(string columnTitle) {
     int result = 0;
 
-    for (int i = 0; i < columnTitle.size(); ++i) {
+    for (size_t i = 0; i < columnTitle.size(); ++i) {
       result *= 26;
       result += columnTitle[i] - 'A' + 1;
     }
