+++ src/zyk.cpp
@@ -8,22 +8,15 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-bool compare(int x, int y)
-{
-    return x <= y;
-}
-int main()
-{
-    vector<int> nums = {2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2};
-    auto compare = [](int x, int y)
-    {
-        return (x < y);
-    };
-    // sort(nums.begin(), nums.end(), compare);
-    sort(nums.begin(), nums.end(), compare);
-    for (int i = 0; i < nums.size(); ++i)
-    {
-        std::cout << nums[i] << " ";
-    }
-    std::cout << std::endl;
+bool compare(int x, int y) { return x <= y; }
+int main() {
+  vector<int> nums = {2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
+                      2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2};
+  auto compare = [](int x, int y) { return (x < y); };
+  // sort(nums.begin(), nums.end(), compare);
+  sort(nums.begin(), nums.end(), compare);
+  for (size_t i = 0; i < nums.size(); ++i) {
+    std::cout << nums[i] << " ";
+  }
+  std::cout << std::endl;
 }
